feat: добавил правило пустой строки после директив и исправил везде yarn lint --fix
