+++ .eslintrc.json
@@ -1,10 +1,11 @@
 {
   "extends": "next/core-web-vitals",
-  "plugins": ["simple-import-sort"],
+  "plugins": ["simple-import-sort", "@stylistic"],
   "rules": {
-     "padding-line-between-statements": [
+    "@stylistic/padding-line-between-statements": [
         "error",
-        [{ "blankLine": "always", "prev": "directive", "next": "*" }, { "blankLine": "any", "prev": "directive", "next": "directive" }]
+        { "blankLine": "always", "prev": "directive", "next": "*" },
+        { "blankLine": "any", "prev": "directive", "next": "directive" }
     ],
     "import/order": [
       "error",
