+++ package.json
@@ -19,6 +19,7 @@
     "redux": "^5.0.1"
   },
   "devDependencies": {
+    "@stylistic/eslint-plugin": "^1.7.0",
     "@types/node": "^20",
     "@types/react": "^18",
     "@types/react-dom": "^18",
