+++ src/app/page.tsx
@@ -1,4 +1,5 @@
 'use client';
+
 import { useState } from 'react';
 
 import { CardsList } from '@/components/CardsList/CardsList';
