+++ src/components/Faq/Faq.tsx
@@ -1,7 +1,9 @@
 import { FaqItem } from '@/components/FaqItem/FaqItem';
-import { testFaqItems } from '@/utils/constants';
+import { testFaqItems } from '@/utils/constans';
 
 import styles from './Faq.module.scss';
+
+
 
 const Faq = () => {
   return (
