+++ src/utils/constans.ts
@@ -69,7 +69,7 @@ const linksItems: LinksItems = {
     },
     {
       text: 'Пример3',
-      id: 2,
+      id: 3,
       linkText: 'Пример4',
       linkUrl: '#',
     },
@@ -95,7 +95,7 @@ const linksItems: LinksItems = {
     },
     {
       text: 'Пример3',
-      id: 2,
+      id: 3,
       linkText: 'Пример8',
       linkUrl: '#',
     },
@@ -121,7 +121,7 @@ const linksItems: LinksItems = {
     },
     {
       text: 'Пример3',
-      id: 2,
+      id: 3,
       linkText: 'Пример12',
       linkUrl: '#',
     },
@@ -141,11 +141,11 @@ const headerLinks: HeaderLink[] = [
 ];
 
 const navLinksMenu: NavLink[] = [
-  { id: 5, name: 'Главная', href: '/' },
-  { id: 1, name: 'Блог', href: '/blog' },
-  { id: 2, name: 'Тренерский состав', href: '/coachs' },
-  { id: 3, name: 'Новости', href: '/news' },
-  { id: 4, name: 'Контакты', href: '/contacts' },
+  { id: 1, name: 'Главная', href: '/' },
+  { id: 2, name: 'Блог', href: '/blog' },
+  { id: 3, name: 'Тренерский состав', href: '/coachs' },
+  { id: 4, name: 'Новости', href: '/news' },
+  { id: 5, name: 'Контакты', href: '/contacts' },
 ];
 
 const blogCards: BlogCard[] = [
@@ -348,4 +348,68 @@ const descriptionPages = [
   },
 ];
 
-export { linksItems, navLinksMenu, buttonsNames, headerLinks, blogCards, coachCards, descriptionPages };
+// тестовый массив с контентом блока FAQ
+const testFaqItems = [
+  {
+    title: 'Пример текста0',
+    text:
+      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста',
+      id: 0
+  },
+  {
+    title: 'Пример текста1',
+    text:
+      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста',
+      id: 1
+  },
+  {
+    title: 'Пример текста2',
+    text:
+      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
+      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
+      'пример текста пример текста пример текста пример текста пример текста',
+      id: 2
+  },
+];
+
+const testCards = [
+  {
+    title: 'Пример текста пример текста пример текста пример текста',
+    subtitle: 'Пример текста',
+    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
+    linkUrl: '#',
+    linkText: 'Пример текста',
+    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
+    id: 0,
+  },
+  {
+    title: 'Пример текста пример текста пример текста пример текста',
+    subtitle: 'Пример текста',
+    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
+    linkUrl: '#',
+    linkText: 'Пример текста',
+    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
+    id: 1,
+  },
+  {
+    title: 'Пример текста пример текста пример текста пример текста',
+    subtitle: 'Пример текста',
+    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
+    linkUrl: '#',
+    linkText: 'Пример текста',
+    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
+    id: 2,
+  },
+];
+
+export { linksItems, navLinksMenu, buttonsNames, headerLinks, blogCards, coachCards, descriptionPages, testFaqItems, testCards };
