--- src/utils/constants.ts
@@ -1,35 +1,0 @@
-// тестовый массив с контентом блока FAQ
-const testFaqItems = [
-  {
-    title: 'Пример текста0',
-    text:
-      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста',
-      id: 0
-  },
-  {
-    title: 'Пример текста1',
-    text:
-      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста',
-      id: 1
-  },
-  {
-    title: 'Пример текста2',
-    text:
-      'Пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста пример текста пример' +
-      'текста пример текста пример текста пример текста пример текста пример текста пример текста' +
-      'пример текста пример текста пример текста пример текста пример текста',
-      id: 2
-  },
-];
-
-export { testFaqItems };
