--- src/utils/constatns.ts
@@ -1,31 +1,0 @@
-const testCards = [
-  {
-    title: 'Пример текста пример текста пример текста пример текста',
-    subtitle: 'Пример текста',
-    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
-    linkUrl: '#',
-    linkText: 'Пример текста',
-    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
-    id: 0,
-  },
-  {
-    title: 'Пример текста пример текста пример текста пример текста',
-    subtitle: 'Пример текста',
-    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
-    linkUrl: '#',
-    linkText: 'Пример текста',
-    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
-    id: 1,
-  },
-  {
-    title: 'Пример текста пример текста пример текста пример текста',
-    subtitle: 'Пример текста',
-    text: 'Пример текста пример текста пример текста пример текста пример текста пример текста пример текста пример',
-    linkUrl: '#',
-    linkText: 'Пример текста',
-    imgUrl: '/sports-teacher-in-gym-class 1.jpg',
-    id: 2,
-  },
-];
-
-export { testCards };
